enhance plan features done from phase 1 to phase6

+++ test/unit/services/ai/multi_ai_service_test.dart
@@ -188,7 +188,7 @@ Best streak: 45 days
       final statsAfter = multiAI.getUsageStats();
       final requestsAfter = statsAfter['total_requests'];
 
-      expect(requestsAfter, greaterThan(requestsBefore));
+      expect(requestsAfter as int, greaterThan(requestsBefore as int));
     });
 
     test('Last used API is tracked', () async {
